03-data-lake-with-google-cloud-storage/examples/uploading-gcs-to-bigquery/pyproject.toml
[tool.poetry]
name = "uploading-gcs-to-bigquery"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "uploading_gcs_to_bigquery"}]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-storage = "^2.8.0"
google-cloud-bigquery = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
